requirements.txt
asgiref==3.10.0
Django==5.2.7
mercadopago==2.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.2
qrcode==8.2
requests==2.33.1
sqlparse==0.5.3
tzdata==2025.2
